config: all: Increase CONFIG_RCU_CPU_STALL_TIMEOUT to 60s (jsc#EB-193)
EB reported RCU stall warnings from xfrm_policy_lookup_bytype for
a workload in bsc#1185675.

CONFIG_RCU_CPU_STALL_TIMEOUT for SLE15-SP2-EB is set to 21s, while
SLE/SLERT runs with CONFIG_RCU_CPU_STALL_TIMEOUT=60
(commit 9e8628bc18). It is possible that 21s leads to false positive
RCU-stall warnings, eg in case a userspace process / custom driver
indeed require that much time in the RCU-read side critical section.

Restore this timeout to 60s for -EB to allow more time for the
workload to exit the RCU-read critical section, with the intention
of avoiding this warning.

Raising the timeout from 21s to 60s does not introduce the risk of
missing fatal RCU stalls, since a stall will persist for >21s if it
is actually fatal. 60s is used as a sane default for SLE/SLERT to
report such stalls.