blacklist.conf:
0c0e37dc1167 x86/ioapic: Force affinity setup before startup
ff363f480e59 x86/msi: Force affinity setup before startup

This whole thing is needed when the affinity change happens after an
interrupt is enabled - in that case an interrupt might get lost. The
magic dance that protects against that is in ...apic/msi.c::msi_set_affinity().

So both would need more involved backport as at least this prerequisite is
needed:

  826da771291f ("genirq: Provide IRQCHIP_AFFINITY_PRE_STARTUP")

which calls irq_setup_affinity(), which, in 4.12 is called
setup_affinity() and static... and there's likely other dependent
changes in-between.

So let's do the backport only when really needed in 12SP5.

Signed-off-by: Borislav Petkov <bp@suse.de>